requirements.txt
APScheduler==3.10.0
asgiref==3.6.0
autopep8==2.0.1
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
defusedxml==0.7.1
Django==4.1.5
django-allauth==0.52.0
django-apscheduler==0.6.2
django-ckeditor==6.5.1
django-filter==22.1
django-js-asset==2.0.0
idna==3.4
oauthlib==3.2.2
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
